Fix: delete data source doesn't work when query results referenced by queries.
