+++ tests/handlers/test_data_sources.py
@@ -1,8 +1,6 @@
-import json
-
 from funcy import pairwise
-
 from tests import BaseTestCase
+
 from redash.models import DataSource, Query
 
 
@@ -77,28 +75,6 @@ class TestDataSourceResourceDelete(BaseTestCase):
     def test_deletes_the_data_source(self):
         data_source = self.factory.create_data_source()
         admin = self.factory.create_admin()
-
-        rv = self.make_request('delete', '/api/data_sources/{}'.format(data_source.id), user=admin)
-
-        self.assertEqual(204, rv.status_code)
-        self.assertIsNone(DataSource.query.get(data_source.id))
-
-    def test_sets_queries_data_source_to_null(self):
-        data_source = self.factory.create_data_source()
-        admin = self.factory.create_admin()
-        query = self.factory.create_query(data_source=data_source)
-
-        rv = self.make_request('delete', '/api/data_sources/{}'.format(data_source.id), user=admin)
-
-        self.assertEqual(204, rv.status_code)
-        self.assertIsNone(DataSource.query.get(data_source.id))
-
-        self.assertIsNone(Query.query.get(query.id).data_source_id)
-
-    def test_deletes_child_models(self):
-        data_source = self.factory.create_data_source()
-        admin = self.factory.create_admin()
-        query_result = self.factory.create_query_result(data_source=data_source)
 
         rv = self.make_request('delete', '/api/data_sources/{}'.format(data_source.id), user=admin)
